adaptive_nav/server/requirements.txt
numpy
openenv-core[core]>=0.2.1
uvicorn[standard]
fastapi
